loadbalancer: Move LBMaps and BPF reconciler to maps / reconciler
