+++ pkg/loadbalancer/maps/test_utils.go
@@ -1,79 +1,24 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package experimental
+package maps
 
 import (
 	"fmt"
 	"net"
-	"net/netip"
 	"slices"
 	"sort"
 	"strconv"
 	"strings"
 
-	"github.com/cilium/statedb"
-	"github.com/cilium/statedb/reconciler"
-
 	"github.com/cilium/cilium/pkg/byteorder"
-	"github.com/cilium/cilium/pkg/clustermesh/types"
 	"github.com/cilium/cilium/pkg/loadbalancer"
-	"github.com/cilium/cilium/pkg/loadbalancer/writer"
 	"github.com/cilium/cilium/pkg/maps/lbmap"
 )
 
-var (
-	// special addresses that are replaced by the test runner.
-	autoAddr = loadbalancer.L3n4Addr{
-		AddrCluster: types.MustParseAddrCluster("0.0.0.1"),
-		L4Addr:      loadbalancer.L4Addr{},
-		Scope:       0,
-	}
-	zeroAddr = loadbalancer.L3n4Addr{
-		AddrCluster: types.MustParseAddrCluster("0.0.0.3"),
-		L4Addr:      loadbalancer.L4Addr{},
-		Scope:       0,
-	}
-
-	extraFrontend = loadbalancer.L3n4Addr{
-		AddrCluster: types.MustParseAddrCluster("10.0.0.2"),
-		L4Addr: loadbalancer.L4Addr{
-			Protocol: loadbalancer.TCP,
-			Port:     80,
-		},
-		Scope: 0,
-	}
-
-	// backend addresses
-	backend1 = loadbalancer.L3n4Addr{
-		AddrCluster: types.MustParseAddrCluster("10.1.0.1"),
-		L4Addr: loadbalancer.L4Addr{
-			Protocol: loadbalancer.TCP,
-			Port:     80,
-		},
-		Scope: 0,
-	}
-	backend2 = loadbalancer.L3n4Addr{
-		AddrCluster: types.MustParseAddrCluster("10.1.0.2"),
-		L4Addr: loadbalancer.L4Addr{
-			Protocol: loadbalancer.TCP,
-			Port:     80,
-		},
-		Scope: 0,
-	}
-
-	// frontendAddrs are assigned to the <auto>/autoAddr. Each test set is run with
-	// each of these.
-	frontendAddrs = []loadbalancer.L3n4Addr{
-		parseAddrPort("10.0.0.1:80"),
-		parseAddrPort("[2001::1]:80"),
-	}
-
-	nodePortAddrs = []netip.Addr{
-		netip.MustParseAddr("10.0.0.3"),
-		netip.MustParseAddr("2002::1"),
-	}
-)
+// MapDump is a dump of a BPF map. These are generated by the dump() method, which
+// solely defines the format.
+type MapDump = string
 
 type numeric interface {
 	~int | ~uint32 | ~uint16
@@ -89,55 +34,6 @@ func sanitizeID[Num numeric](n Num, sanitize bool) string {
 		return "<zero>"
 	}
 	return "<non-zero>"
-}
-
-func parseAddrPort(s string) loadbalancer.L3n4Addr {
-	addrS, portS, found := strings.Cut(s, "]:")
-	if found {
-		// IPv6
-		addrS = addrS[1:] // drop [
-	} else {
-		// IPv4
-		addrS, portS, found = strings.Cut(s, ":")
-		if !found {
-			panic("bad <ip:port>")
-		}
-	}
-	addr := types.MustParseAddrCluster(addrS)
-	port, _ := strconv.ParseInt(portS, 10, 16)
-	return *loadbalancer.NewL3n4Addr(
-		loadbalancer.TCP,
-		addr, uint16(port), loadbalancer.ScopeExternal,
-	)
-
-}
-
-// MapDump is a dump of a BPF map. These are generated by the dump() method, which
-// solely defines the format.
-type MapDump = string
-
-func dumpLBMapsWithReplace(lbmaps LBMaps, feAddr loadbalancer.L3n4Addr, sanitizeIDs bool) (out []MapDump) {
-	replaceAddr := func(addr net.IP, port uint16) (s string) {
-		s = addr.String()
-		if addr.To4() == nil {
-			s = "[" + s + "]"
-		}
-		s = fmt.Sprintf("%s:%d", s, port)
-		if addr.IsUnspecified() {
-			s = "<zero>"
-			return
-		}
-		switch addr.String() {
-		case feAddr.AddrCluster.String():
-			s = "<auto>"
-		case nodePortAddrs[0].String():
-			s = "<nodePort>"
-		case nodePortAddrs[1].String():
-			s = "<nodePort>"
-		}
-		return
-	}
-	return DumpLBMaps(lbmaps, sanitizeIDs, replaceAddr)
 }
 
 // DumpLBMaps the load-balancing maps into a concise format for assertions in tests.
@@ -293,30 +189,4 @@ func DumpLBMaps(lbmaps LBMaps, sanitizeIDs bool, customizeAddr func(net.IP, uint
 
 	sort.Strings(out)
 	return
-}
-
-func FastCheckTables(db *statedb.DB, writer *writer.Writer, expectedFrontends int, lastPendingRevision statedb.Revision) (reconciled bool, nextRevision statedb.Revision) {
-	txn := db.ReadTxn()
-	if writer.Frontends().NumObjects(txn) < expectedFrontends {
-		return false, 0
-	}
-	var rev uint64
-	var fe *loadbalancer.Frontend
-	for fe, rev = range writer.Frontends().LowerBound(txn, statedb.ByRevision[*loadbalancer.Frontend](lastPendingRevision)) {
-		if fe.Status.Kind != reconciler.StatusKindDone {
-			return false, rev
-		}
-	}
-	return true, rev // Here, it is the last reconciled revision rather than the first non-reconciled revision.
-}
-
-func FastCheckEmptyTablesAndState(db *statedb.DB, writer *writer.Writer, bo *BPFOps) bool {
-	txn := db.ReadTxn()
-	if writer.Frontends().NumObjects(txn) > 0 || writer.Backends().NumObjects(txn) > 0 || writer.Services().NumObjects(txn) > 0 {
-		return false
-	}
-	if len(bo.backendReferences) > 0 || len(bo.backendStates) > 0 || len(bo.nodePortAddrByPort) > 0 || len(bo.serviceIDAlloc.entities) > 0 || len(bo.backendIDAlloc.entities) > 0 {
-		return false
-	}
-	return true
 }
